Implemented Hexagon Test

+++ src/model/HexagonModel.java
@@ -12,9 +12,10 @@ public class HexagonModel extends ShapeModel2D {
         // Calculate circumradius of the hexagon using Pythagoras
         final int a = endX - startX;
         final int b = endY - startY;
-        final int radius = (int) (0.5  * Math.sqrt(Math.pow(a, 2) + Math.pow(b, 2)));
+        final double radius = 0.5  * Math.sqrt(Math.pow(a, 2) + Math.pow(b, 2));
 
-        // Calculate center point as 1/2 distance between start and end
+        // Calculate center point as 1/2 distance between start and end (must be integer
+        // as it represents an actual point.
         final int midX = (int) (startX + 0.5 * a);
         final int midY = (int) (startY + 0.5 * b);
 
@@ -24,9 +25,8 @@ public class HexagonModel extends ShapeModel2D {
         final double degMultiple = Math.toRadians(60);
 
         for (int i = 0; i < xpoints.length; i++) {
-
-            xpoints[i] = (int) (radius * Math.cos((- degMultiple) * i) + midX);
-            ypoints[i] = (int) (radius * Math.sin((- degMultiple) * i) + midY);
+             xpoints[i] = (int) (radius * Math.cos((- degMultiple) * i) + midX);
+             ypoints[i] = (int) (radius * Math.sin((- degMultiple) * i) + midY);
         }
 
         this.xpoints = xpoints;
